+++ lib/httparty/utils.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module HTTParty
   module Utils
     def self.stringify_keys(hash)